+++ addons/l10n_et_payroll/tests/__init__.py
@@ -1,3 +1,4 @@
+from . import test_config_seeding
 from . import test_payroll_compute
 from . import test_payroll_run
 from . import test_payroll_reports